update tests and seperate them
seperate the tests between converter and styles
